Desktop/Syra MOl/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
motor==3.6.0
dnspython==2.6.1
requests==2.32.3
scikit-learn==1.5.2
jinja2==3.1.4
flask
pyttsx3
